+++ src/components/atoms/ImageInput/style.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const ImageInputButton = styled.button`
+export const Button = styled.button`
   width: 150px;
   height: 50px;
 `;
