+++ src/components/organisms/FileList/style.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const FileListDiv = styled.div`
+export const Container = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: space-between;
