+++ src/pages/index.js
@@ -1,2 +1,1 @@
-export { Example1Page } from './Example1';
 export { Example2Page } from './Example2';
